Rename blocks and move frontend code to blocks
