+++ src/blocks/block-hydration-experiments-child/index.js
@@ -1,10 +1,10 @@
-import Text from '../../frontend/text';
 import { registerBlockType } from '../../gutenberg-packages/wordpress-blocks';
 import Edit from './edit';
+import Frontend from './frontend';
 import './style.scss';
 
 registerBlockType( 'bhe/block-hydration-experiments-child', {
 	edit: Edit,
 	// The Save component is derived from the Frontend component.
-	frontend: Text,
+	frontend: Frontend,
 } );
