+++ src/blocks/block-hydration-experiments-child/view.js
@@ -1,4 +1,4 @@
-import Text from '../../frontend/text';
 import { registerBlockType } from '../../gutenberg-packages/frontend';
+import Frontend from './frontend';
 
-registerBlockType( 'bhe/block-hydration-experiments-child', Text );
+registerBlockType( 'bhe/block-hydration-experiments-child', Frontend );
