+++ src/blocks/block-hydration-experiments-parent/frontend/index.js
@@ -1,8 +1,8 @@
-import { useState } from '../gutenberg-packages/wordpress-element';
+import { useState } from '../../../gutenberg-packages/wordpress-element';
 import Title from '../shared/title';
 import Button from './button';
 
-const Block = ( { blockProps, attributes, children } ) => {
+const Frontend = ( { blockProps, attributes, children } ) => {
 	const [ show, setShow ] = useState( false );
 	const [ counter, setCounter ] = useState( 0 );
 
@@ -16,4 +16,4 @@ const Block = ( { blockProps, attributes, children } ) => {
 	);
 };
 
-export default Block;
+export default Frontend;
