+++ src/blocks/block-hydration-experiments-parent/shared/title.js
@@ -1,4 +1,4 @@
-import { RichText } from '../gutenberg-packages/wordpress-blockeditor';
+import { RichText } from '../../../gutenberg-packages/wordpress-blockeditor';
 
 const Title = ( { message, ...props } ) => (
 	<RichText tagName='h2' className='title' {...props}>{message}</RichText>
