+++ src/blocks/block-hydration-experiments-parent/view.js
@@ -1,4 +1,4 @@
-import Frontend from '../../frontend';
 import { registerBlockType } from '../../gutenberg-packages/frontend';
+import Frontend from './frontend';
 
 registerBlockType( 'bhe/block-hydration-experiments-parent', Frontend );
